microservices POC
