+++ libs/README.md
@@ -1,0 +1,3 @@
+# Spark core functionality js libraries
+
+Provides general core functionality like mail, logging, db, etc..
