+++ libs/testlib.js
@@ -1,0 +1,9 @@
+const testlib = module.exports = {};
+
+testlib.getAdminHttpRequest = function() {
+    return {
+        user: {
+            isAdmin: true
+        }
+    };
+};
