+++ modules/admin/libs/admin.js
@@ -2,7 +2,7 @@
  * provides functionality related to the admin UI
  */
 var modules = require('../../../libs/modules');
-var userRole = modules.require('core', 'libs/user_role');
+var userRole = modules.require('users', 'libs/user_role');
 
 /**
  * constants
@@ -70,11 +70,11 @@ var datatableAdmin = function(name, router, opts) {
                 "lengthMenu": PAGE_LENGTH_MENU, // the options in the page length select list
                 "pageLength": DEFAULT_PAGE_LENGTH, // initial page length (number of rows per page)
                 "order": opts.defaultOrder,
-                "addUrl": ADMIN_URL_PREFIX + ADD_URL_TEMPLATE.replace('{name}', name),
                 "editKey": opts.editKey,
                 "editUrl": ADMIN_URL_PREFIX + EDIT_URL_TEMPLATE.replace('{name}', name)
             }),
-            addTitle: opts.addTitle
+            addTitle: opts.addTitle,
+            addUrl: ADMIN_URL_PREFIX + ADD_URL_TEMPLATE.replace('{name}', name),
         });
     };
 
@@ -91,13 +91,10 @@ var datatableAdmin = function(name, router, opts) {
         };
         if (req.params.object_id) {
             // edit existing object - need to fetch it first
-            var fetchObj = {};
-            fetchObj[opts.editKey] = req.params.object_id;
-            api.core.users.fetch(fetchObj).then(function(user) {
+            opts.api.fetchById(req, req.params.object_id).then(function(object) {
                 opts.columns.forEach(function(column) {
-                    // if (object.attributes[column.attr]) {
-                    if (user[column.attr]) {
-                        column.value = user[column.attr];
+                    if (object[column.attr]) {
+                        column.value = object[column.attr];
                     } else if (req.body && req.body[column.attr]) {
                         column.value = req.body[column.attr]
                     } else {
@@ -106,11 +103,6 @@ var datatableAdmin = function(name, router, opts) {
                 });
                 _adminRender(opts.editTitle);
             });
-            // var forgeObj={};
-            // forgeObj[opts.editKey] = req.params.object_id;
-            // opts.model.forge(forgeObj).fetch().then(function(object) {
-            //
-            // });
         } else {
             // add new object
             opts.columns.forEach(function(column) {
@@ -124,7 +116,7 @@ var datatableAdmin = function(name, router, opts) {
         }
     };
 
-    router.get('/' + name + '/table', userRole.isAdmin(), function(req, res) {
+    router.get('/admin/' + name + '/table', userRole.isAdmin(), function(req, res) {
         try {
             // initialize from datatables query params
             var resultsPerPage = req.query.length;
@@ -140,50 +132,42 @@ var datatableAdmin = function(name, router, opts) {
             });
         }
         var recordsTotal = 0;
-        opts.model
-            .count()
-            .then(function(count) {
-                recordsTotal = count;
-                opts.model.query(function(qb) {
-                        if (searchTerm) {
-                            opts.filter(qb, searchTerm);
-                        }
-                    })
-                    .orderBy(orderBy)
-                    .fetchPage({
-                        columns: opts.selectColumns,
-                        pageSize: resultsPerPage,
-                        page: currentPage
-                    })
-                    .then(function(rows) {
-                        // rows.pagination: { page: 1, pageSize: 2, rowCount: 5, pageCount: 3 }
-                        var data = rows.toJSON();
-                        data.forEach(function(row) {
-                            row.actions = "";
-                        });
-                        res.status(200).json({
-                            data: data,
-                            recordsTotal: recordsTotal, // total records before filtering
-                            recordsFiltered: rows.pagination.rowCount // total records after filtering
-                        })
-                    });
-            })
-            .catch(function(e) {
-                res.status(500).json({
-                    error: true,
-                    data: {
-                        message: e.message
-                    }
+        opts.api.getTotalCount(req).then(function(count) {
+            recordsTotal = count;
+            opts.api.fetchPage(req, {
+                searchTerm: searchTerm,
+                orderBy: orderBy,
+                columns: opts.selectColumns,
+                pageSize: resultsPerPage,
+                page: currentPage
+            }).then(function(fetchResponse) {
+                // fetchResponse.pagination: { page: 1, pageSize: 2, rowCount: 5, pageCount: 3 }
+                // fetchResponse.rows = [{}, {}..]
+                fetchResponse.rows.forEach(function(row) {
+                    row.actions = "";
                 });
+                res.status(200).json({
+                    data: fetchResponse.rows,
+                    recordsTotal: recordsTotal, // total records before filtering
+                    recordsFiltered: fetchResponse.pagination.rowCount // total records after filtering
+                })
             });
+        }).catch(function(e) {
+            res.status(500).json({
+                error: true,
+                data: {
+                    message: e.message
+                }
+            });
+        });
     });
 
-    router.get('/' + name + '/add', userRole.isAdmin(), function(req, res) {
+    router.get('/admin/' + name + '/add', userRole.isAdmin(), function(req, res) {
         _renderEditPage(req, res, "")
     });
 
-    router.post('/' + name + '/add', userRole.isAdmin(), function(req, res) {
-        opts.addCallback(req.body, function(ok, msg) {
+    router.post('/admin/' + name + '/add', userRole.isAdmin(), function(req, res) {
+        opts.addCallback(req, req.body, function(ok, msg) {
             if (ok) {
                 _renderTablePage(req, res, msg);
             } else {
@@ -192,12 +176,12 @@ var datatableAdmin = function(name, router, opts) {
         });
     });
 
-    router.get('/' + name + '/edit/:object_id', userRole.isAdmin(), function(req, res) {
+    router.get('/admin/' + name + '/edit/:object_id', userRole.isAdmin(), function(req, res) {
         _renderEditPage(req, res, "")
     });
 
-    router.post('/' + name + '/edit/:object_id', userRole.isAdmin(), function(req, res) {
-        opts.editCallback(req.params.object_id, req.body, function(ok, msg) {
+    router.post('/admin/' + name + '/edit/:object_id', userRole.isAdmin(), function(req, res) {
+        opts.editCallback(req, req.params.object_id, req.body, function(ok, msg) {
             if (ok) {
                 _renderEditPage(req, res, false, msg);
             } else {
@@ -206,7 +190,7 @@ var datatableAdmin = function(name, router, opts) {
         })
     });
 
-    router.get('/' + name, userRole.isAdmin(), function(req, res) {
+    router.get('/admin/' + name, userRole.isAdmin(), function(req, res) {
         _renderTablePage(req, res, "")
     });
