+++ modules/payment/README.md
@@ -1,0 +1,3 @@
+# Spark payment module
+
+Provides payment functionality
