+++ modules/payment/libs/payment.js
@@ -1,6 +1,6 @@
 var request = require('request');
 var Payment = require('../models/payment').Payment;
-var log = require('./logger.js')(module);
+var log = require('./../../../libs/logger.js')(module);
 
 var config = require('config');
 var paymentConfig = config.get('payment');
