+++ modules/payment/models/payment.js
@@ -1,5 +1,5 @@
-var bookshelf = require('../libs/db').bookshelf;
-var constants = require('./constants.js');
+var bookshelf = require('../../../libs/db').bookshelf;
+var constants = require('./../../../models/constants.js');
 
 var Payment = bookshelf.Model.extend({
     tableName: constants.PAYMENTS_TABLE_NAME,
